TFP-1302 rydder vekk løpende medlemskap toggle
