+++ packages/fp-behandling-forstegang-og-revurdering/src/behandlingsprosess/definition/foreldrepengerBpDefinition.jsx
@@ -1,5 +1,5 @@
 import { BehandlingspunktProperties } from '@fpsak-frontend/fp-behandling-felles';
-import { behandlingspunktCodes as bpc, featureToggle } from '@fpsak-frontend/fp-felles';
+import { behandlingspunktCodes as bpc } from '@fpsak-frontend/fp-felles';
 import ac from '@fpsak-frontend/kodeverk/src/aksjonspunktCodes';
 import vt from '@fpsak-frontend/kodeverk/src/vilkarType';
 import bt from '@fpsak-frontend/kodeverk/src/behandlingType';
@@ -46,7 +46,6 @@ const getStatusFromResultatstruktur = ({ resultatstruktur, uttaksresultat }) => 
   return vut.IKKE_VURDERT;
 };
 
-const behandlingTypeEquals = (...behandlingTypes) => ({ behandlingType }) => behandlingTypes.some((b) => b === behandlingType.kode);
 const hasNonDefaultBehandlingspunkt = (builderData, bpLength) => bpLength > 0;
 
 const isNotRevurderingAndManualOpplysningspliktAp = ({ behandlingType, aksjonspunkter }) => {
@@ -55,7 +54,6 @@ const isNotRevurderingAndManualOpplysningspliktAp = ({ behandlingType, aksjonspu
   return !(isRevurdering && !hasAp);
 };
 
-const hasLøpendeMedlemskapOn = ({ featureToggles }) => featureToggles[featureToggle.LØPENDE_MEDLESMKAP];
 /**
  * Rekkefølgen i listene under bestemmer behandlingspunkt-rekkefølgen i GUI.
  * @see BehandlingspunktProperties.Builder for mer informasjon.
@@ -112,8 +110,7 @@ const foreldrepengerBuilders = [
 
   new BehandlingspunktProperties.Builder(bpc.FORTSATTMEDLEMSKAP, 'FortsattMedlemskap')
     .withVilkarTypes(vt.MEDLEMSKAPSVILKÅRET_LØPENDE)
-    .withAksjonspunktCodes(ac.OVERSTYR_LØPENDE_MEDLEMSKAPSVILKAR)
-    .withVisibilityWhen(behandlingTypeEquals(bt.REVURDERING), hasLøpendeMedlemskapOn),
+    .withAksjonspunktCodes(ac.OVERSTYR_LØPENDE_MEDLEMSKAPSVILKAR),
 
   new BehandlingspunktProperties.Builder(bpc.UTTAK, 'Uttak')
     .withVisibilityWhen(hasNonDefaultBehandlingspunkt)
