+++ packages/fp-behandling-forstegang-og-revurdering/src/fakta/components/medlemskap/oppholdInntektOgPerioder/OppholdINorgeOgAdresserFaktaPanel.spec.jsx
@@ -7,7 +7,7 @@ import aksjonspunktStatus from '@fpsak-frontend/kodeverk/src/aksjonspunktStatus'
 import aksjonspunktCodes from '@fpsak-frontend/kodeverk/src/aksjonspunktCodes';
 import opplysningAdresseType from '@fpsak-frontend/kodeverk/src/opplysningAdresseType';
 import { Normaltekst } from 'nav-frontend-typografi';
-import { RadioOption } from '@fpsak-frontend/form';
+import { FormattedMessage } from 'react-intl';
 import OppholdINorgeOgAdresserFaktaPanel from './OppholdINorgeOgAdresserFaktaPanel';
 
 describe('<OppholdINorgeOgAdresserFaktaPanel>', () => {
@@ -39,24 +39,24 @@ describe('<OppholdINorgeOgAdresserFaktaPanel>', () => {
     const wrapper = shallowWithIntl(<OppholdINorgeOgAdresserFaktaPanel.WrappedComponent
       intl={intlMock}
       readOnly={false}
-      hasAksjonspunkt={false}
+      hasBosattAksjonspunkt={false}
       isBosattAksjonspunktClosed={false}
       opphold={opphold}
       foreldre={foreldre}
     />);
     const felter = wrapper.find(Normaltekst);
     expect(felter).to.have.length(5);
-    expect(felter.first().childAt(0).text()).to.eql('Ja');
-    expect(felter.at(1).childAt(0).text()).to.eql('Ja');
+    expect(felter.first().childAt(0).find(FormattedMessage).prop('id')).to.eql('OppholdINorgeOgAdresserFaktaPanel.Yes');
+    expect(felter.at(1).childAt(0).find(FormattedMessage).prop('id')).to.eql('OppholdINorgeOgAdresserFaktaPanel.Yes');
     expect(felter.at(2).childAt(0).text()).to.eql('Sverige');
-    expect(felter.at(4).childAt(0).text()).to.eql('Nei');
+    expect(felter.at(4).childAt(0).find(FormattedMessage).prop('id')).to.eql('OppholdINorgeOgAdresserFaktaPanel.No');
   });
 
   it('skal rendre form som viser bosatt informasjon', () => {
     const wrapper = shallowWithIntl(<OppholdINorgeOgAdresserFaktaPanel.WrappedComponent
       intl={intlMock}
       readOnly={false}
-      hasAksjonspunkt={false}
+      hasBosattAksjonspunkt={false}
       isBosattAksjonspunktClosed={false}
       opphold={opphold}
       foreldre={foreldre}
@@ -81,64 +81,62 @@ describe('<OppholdINorgeOgAdresserFaktaPanel>', () => {
     const wrapper = shallowWithIntl(<OppholdINorgeOgAdresserFaktaPanel.WrappedComponent
       intl={intlMock}
       readOnly={false}
-      hasAksjonspunkt
+      hasBosattAksjonspunkt
       isBosattAksjonspunktClosed={false}
       opphold={opphold}
       foreldre={toForeldre}
     />);
 
-    const radioFields = wrapper.find(RadioOption);
+    const radioFields = wrapper.find('RadioOption');
     expect(radioFields).to.have.length(2);
     expect(radioFields.first().prop('label').id).to.eql('OppholdINorgeOgAdresserFaktaPanel.ResidingInNorway');
   });
 
   it('skal sette initielle verdier', () => {
-    const medlem = {
+    const periode = {
+      aksjonspunkter: [aksjonspunktCodes.AVKLAR_OM_BRUKER_ER_BOSATT],
       bosattVurdering: true,
-    };
-    const soknad = {
-      oppgittTilknytning: opphold,
-    };
-    const personopplysning = {
-      navn: 'Espen Utvikler',
-      personstatus: {
-        personstatus: {
-          kode: 'UTVANDRET',
-          navn: 'Utvandret',
-        },
-      },
-      avklartPersonstatus: {
-        overstyrtPersonstatus: {
-          kode: personstatusType.BOSATT,
-          navn: 'Bosatt',
-        },
-      },
-      adresser: [{
-        adresselinje1: 'Vei 1',
-        postNummer: '1000',
-        poststed: 'Oslo',
-        opplysningAdresseType: {
-          kode: opplysningAdresseType.POSTADRESSE,
-          navn: 'Bostedsadresse',
-        },
-      }],
-      annenPart: {
-        navn: 'Petra Tester',
+      personopplysninger: {
+        navn: 'Espen Utvikler',
         personstatus: {
           personstatus: {
             kode: 'UTVANDRET',
             navn: 'Utvandret',
           },
         },
+        avklartPersonstatus: {
+          overstyrtPersonstatus: {
+            kode: personstatusType.BOSATT,
+            navn: 'Bosatt',
+          },
+        },
         adresser: [{
-          adresselinje1: 'Vei 2',
-          postNummer: '2000',
-          poststed: 'Stockholm',
+          adresselinje1: 'Vei 1',
+          postNummer: '1000',
+          poststed: 'Oslo',
           opplysningAdresseType: {
-            kode: opplysningAdresseType.UTENLANDSK_POSTADRESSE,
+            kode: opplysningAdresseType.POSTADRESSE,
             navn: 'Bostedsadresse',
           },
         }],
+        annenPart: {
+          navn: 'Petra Tester',
+          personstatus: {
+            personstatus: {
+              kode: 'UTVANDRET',
+              navn: 'Utvandret',
+            },
+          },
+          adresser: [{
+            adresselinje1: 'Vei 2',
+            postNummer: '2000',
+            poststed: 'Stockholm',
+            opplysningAdresseType: {
+              kode: opplysningAdresseType.UTENLANDSK_POSTADRESSE,
+              navn: 'Bostedsadresse',
+            },
+          }],
+        },
       },
     };
 
@@ -150,8 +148,11 @@ describe('<OppholdINorgeOgAdresserFaktaPanel>', () => {
         kode: aksjonspunktStatus.OPPRETTET,
       },
     }];
+    const soknad = {
+      oppgittTilknytning: opphold,
+    };
 
-    const initialValues = OppholdINorgeOgAdresserFaktaPanel.buildInitialValues(soknad, personopplysning, medlem, aksjonspunkter);
+    const initialValues = OppholdINorgeOgAdresserFaktaPanel.buildInitialValues(soknad, periode, aksjonspunkter);
 
     expect(initialValues).to.eql({
       foreldre: [{
