+++ packages/fp-behandling-forstegang-og-revurdering/src/fakta/components/medlemskap/oppholdInntektOgPerioder/OppholdInntektOgPerioderForm.jsx
@@ -1,37 +1,43 @@
-import React from 'react';
+import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import { createSelector } from 'reselect';
-import { FormattedMessage } from 'react-intl';
-import { formPropTypes } from 'redux-form';
 import { connect } from 'react-redux';
+import { createSelector } from 'reselect';
+import { change as reduxFormChange, formPropTypes, reset as reduxFormReset } from 'redux-form';
+import { FormattedMessage, injectIntl } from 'react-intl';
+import { bindActionCreators } from 'redux';
+import { Hovedknapp } from 'nav-frontend-knapper';
 
-import { AksjonspunktHelpText, ElementWrapper, VerticalSpacer } from '@fpsak-frontend/shared-components';
-import { getKodeverknavnFn, FaktaBegrunnelseTextField } from '@fpsak-frontend/fp-felles';
+import { AksjonspunktHelpText, VerticalSpacer } from '@fpsak-frontend/shared-components';
 import { isAksjonspunktOpen } from '@fpsak-frontend/kodeverk/src/aksjonspunktStatus';
 import aksjonspunktCodes, { hasAksjonspunkt } from '@fpsak-frontend/kodeverk/src/aksjonspunktCodes';
+import { aksjonspunktPropType } from '@fpsak-frontend/prop-types';
+import { getBehandlingFormPrefix } from '@fpsak-frontend/fp-felles';
+import { guid } from '@fpsak-frontend/utils';
+
 import {
-  getBehandlingMedlem,
+  getBehandlingMedlemNew,
   getBehandlingRevurderingAvFortsattMedlemskapFom,
-  getPersonopplysning,
+  isBehandlingRevurderingFortsattMedlemskap,
 } from 'behandlingForstegangOgRevurdering/src/behandlingSelectors';
 import behandlingSelectors from 'behandlingForstegangOgRevurdering/src/selectors/forsteOgRevBehandlingSelectors';
-import { behandlingFormForstegangOgRevurdering } from 'behandlingForstegangOgRevurdering/src/behandlingFormForstegangOgRevurdering';
-import FaktaSubmitButton from 'behandlingForstegangOgRevurdering/src/fakta/components/FaktaSubmitButton';
-import { aksjonspunktPropType } from '@fpsak-frontend/prop-types';
-import { getAlleKodeverk, getFagsakPerson, getKodeverk } from 'behandlingForstegangOgRevurdering/src/duckBehandlingForstegangOgRev';
-import kodeverkTyper from '@fpsak-frontend/kodeverk/src/kodeverkTyper';
-import OppholdINorgeOgAdresserFaktaPanel from './OppholdINorgeOgAdresserFaktaPanel';
-import InntektOgYtelserFaktaPanel from './InntektOgYtelserFaktaPanel';
-import PerioderMedMedlemskapFaktaPanel from './PerioderMedMedlemskapFaktaPanel';
-import StatusForBorgerFaktaPanel from './StatusForBorgerFaktaPanel';
-import FortsattMedlemskapFaktaPanel from './FortsattMedlemskapFaktaPanel';
+import {
+  behandlingFormForstegangOgRevurdering,
+  behandlingFormValueSelector,
+} from 'behandlingForstegangOgRevurdering/src/behandlingFormForstegangOgRevurdering';
+import { getFagsakPerson, getSelectedBehandlingId } from 'behandlingForstegangOgRevurdering/src/duckBehandlingForstegangOgRev';
+import OppholdInntektOgPeriodeForm from './OppholdInntektOgPeriodeForm';
+import MedlemskapEndringerTabell from './MedlemskapEndringerTabell';
+
+/**
+ * OppholdInntektOgPerioderForm
+ *
+ * Presentasjonskomponent. Har ansvar for å sette opp Redux Formen for faktapenelet til Medlemskapsvilkåret.
+ */
 
 const {
   AVKLAR_STARTDATO_FOR_FORELDREPENGERPERIODEN, AVKLAR_OM_BRUKER_ER_BOSATT,
   AVKLAR_OM_BRUKER_HAR_GYLDIG_PERIODE, AVKLAR_OPPHOLDSRETT, AVKLAR_LOVLIG_OPPHOLD, AVKLAR_FORTSATT_MEDLEMSKAP,
 } = aksjonspunktCodes;
-
-const shouldSubmitAksjonspunkt = (aksjonspunktCode, aksjonspunkter) => aksjonspunkter.some((ap) => ap.definisjon.kode === aksjonspunktCode && ap.erAktivt);
 
 const getHelpTexts = (aksjonspunkter) => {
   const helpTexts = [];
@@ -53,112 +59,231 @@ const getHelpTexts = (aksjonspunkter) => {
   return helpTexts;
 };
 
-/**
- * OppholdInntektOgPerioderForm
- *
- * Presentasjonskomponent. Har ansvar for å sette opp Redux Formen for faktapenelet til Medlemskapsvilkåret.
- */
-export const OppholdInntektOgPerioderForm = ({
-  hasOpenAksjonspunkter,
-  submittable,
-  aksjonspunkter,
-  readOnly,
-  initialValues,
-  ...formProps
-}) => (
-  <form onSubmit={formProps.handleSubmit}>
-    <AksjonspunktHelpText isAksjonspunktOpen={hasOpenAksjonspunkter || !submittable}>{getHelpTexts(aksjonspunkter)}</AksjonspunktHelpText>
-    { hasAksjonspunkt(AVKLAR_FORTSATT_MEDLEMSKAP, aksjonspunkter)
-      && <FortsattMedlemskapFaktaPanel readOnly={readOnly} />}
-    <OppholdINorgeOgAdresserFaktaPanel readOnly={readOnly} />
-    <InntektOgYtelserFaktaPanel />
-    <PerioderMedMedlemskapFaktaPanel readOnly={readOnly} />
-    { (hasAksjonspunkt(AVKLAR_OPPHOLDSRETT, aksjonspunkter) || hasAksjonspunkt(AVKLAR_LOVLIG_OPPHOLD, aksjonspunkter))
-      && <StatusForBorgerFaktaPanel readOnly={readOnly} />}
-    { aksjonspunkter && aksjonspunkter.length > 0
-      && (
-      <ElementWrapper>
-        <VerticalSpacer twentyPx />
-        <FaktaBegrunnelseTextField isDirty={formProps.dirty} isSubmittable={submittable} isReadOnly={readOnly} hasBegrunnelse={!!initialValues.begrunnelse} />
-        <VerticalSpacer twentyPx />
-        <FaktaSubmitButton formName={formProps.form} isSubmittable={submittable} isReadOnly={readOnly} hasOpenAksjonspunkter={hasOpenAksjonspunkter} />
-      </ElementWrapper>
-      )}
-  </form>
-);
+const createNewPerioder = (perioder, id, values) => {
+  const updatedIndex = perioder.findIndex((p) => p.id === id);
+  const updatedPeriode = perioder.find((p) => p.id === id);
 
-OppholdInntektOgPerioderForm.propTypes = {
+  return [
+    ...perioder.slice(0, updatedIndex),
+    {
+      ...updatedPeriode,
+      ...values,
+    },
+    ...perioder.slice(updatedIndex + 1),
+  ];
+};
+
+export class OppholdInntektOgPerioderFormNew extends Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      valgtPeriode: undefined,
+    };
+
+    this.velgPeriodeCallback = this.velgPeriodeCallback.bind(this);
+    this.updateOppholdInntektPeriode = this.updateOppholdInntektPeriode.bind(this);
+    this.isConfirmButtonDisabled = this.isConfirmButtonDisabled.bind(this);
+    this.periodeResetCallback = this.periodeResetCallback.bind(this);
+  }
+
+
+  componentWillMount() {
+    const { initialValues } = this.props;
+    const defaultPeriode = initialValues.perioder ? initialValues.perioder[0] : {};
+    this.setValgtPeriode(defaultPeriode);
+  }
+
+  setValgtPeriode(valgtPeriode) {
+    if (!valgtPeriode) {
+      const { initialValues } = this.props;
+      const defaultPeriode = initialValues.perioder ? initialValues.perioder[0] : {};
+      this.setState({ valgtPeriode: defaultPeriode });
+    }
+    this.setState({ valgtPeriode });
+  }
+
+  periodeResetCallback() {
+    const { behandlingFormPrefix, reduxFormReset: formReset } = this.props;
+    const { valgtPeriode } = this.state;
+    if (valgtPeriode) {
+      formReset(`${behandlingFormPrefix}.OppholdInntektOgPeriodeForm-${valgtPeriode.id}`);
+    }
+  }
+
+  velgPeriodeCallback(p, id, periode) {
+    const valgtPeriode = {
+      id,
+      ...periode,
+    };
+    this.setState({ valgtPeriode });
+  }
+
+  updateOppholdInntektPeriode(values) {
+    const {
+      behandlingFormPrefix, perioder, reduxFormChange: formChange,
+    } = this.props;
+
+    const updatedPeriode = perioder.find((p) => p.id === values.id);
+
+    const newPeriodeObject = {
+      ...updatedPeriode,
+      ...values,
+    };
+
+    const newPerioder = createNewPerioder(perioder, values.id, newPeriodeObject);
+
+    formChange(`${behandlingFormPrefix}.OppholdInntektOgPerioderForm`, 'perioder', newPerioder);
+  }
+
+  isConfirmButtonDisabled() {
+    const {
+      perioder, readOnly, submitting, dirty,
+    } = this.props;
+
+    if (!dirty) {
+      return true;
+    }
+
+    if (perioder && perioder.length > 0) {
+      const ubekreftPerioder = perioder.filter((periode) => periode.aksjonspunkter.length > 0 && periode.begrunnelse === null);
+
+      if (ubekreftPerioder.length > 0) {
+        return true;
+      }
+    }
+
+    return submitting || readOnly;
+  }
+
+  render() {
+    const {
+      hasOpenAksjonspunkter,
+      submittable,
+      aksjonspunkter,
+      readOnly,
+      submitting,
+      isRevurdering,
+      ...formProps
+    } = this.props;
+
+    const { valgtPeriode } = this.state;
+
+    return (
+      <form onSubmit={formProps.handleSubmit}>
+        <AksjonspunktHelpText isAksjonspunktOpen={hasOpenAksjonspunkter || !submittable}>{getHelpTexts(aksjonspunkter)}</AksjonspunktHelpText>
+
+        { hasAksjonspunkt(AVKLAR_FORTSATT_MEDLEMSKAP, aksjonspunkter) && (
+        <MedlemskapEndringerTabell
+          selectedId={valgtPeriode ? valgtPeriode.id : undefined}
+          velgPeriodeCallback={this.velgPeriodeCallback}
+        />
+        )}
+
+        {valgtPeriode && (
+        <OppholdInntektOgPeriodeForm
+          isRevurdering={isRevurdering}
+          readOnly={readOnly}
+          valgtPeriode={valgtPeriode}
+          aksjonspunkter={aksjonspunkter}
+          submittable={submittable}
+          updateOppholdInntektPeriode={this.updateOppholdInntektPeriode}
+          periodeResetCallback={this.periodeResetCallback}
+        />
+        )}
+
+        <VerticalSpacer twentyPx />
+        <Hovedknapp
+          mini
+          disabled={this.isConfirmButtonDisabled()}
+          spinner={submitting}
+        >
+          <FormattedMessage id="OppholdInntektOgPerioder.Bekreft" />
+        </Hovedknapp>
+      </form>
+    );
+  }
+}
+
+OppholdInntektOgPerioderFormNew.propTypes = {
+  intl: PropTypes.shape().isRequired,
   hasOpenAksjonspunkter: PropTypes.bool.isRequired,
   submittable: PropTypes.bool.isRequired,
   aksjonspunkter: PropTypes.arrayOf(aksjonspunktPropType.isRequired).isRequired,
   readOnly: PropTypes.bool.isRequired,
+  submitting: PropTypes.bool.isRequired,
+  isRevurdering: PropTypes.bool.isRequired,
   ...formPropTypes,
 };
 
 const medlemAksjonspunkter = [AVKLAR_STARTDATO_FOR_FORELDREPENGERPERIODEN, AVKLAR_OM_BRUKER_ER_BOSATT, AVKLAR_OM_BRUKER_HAR_GYLDIG_PERIODE,
-  AVKLAR_OPPHOLDSRETT, AVKLAR_LOVLIG_OPPHOLD];
+  AVKLAR_OPPHOLDSRETT, AVKLAR_LOVLIG_OPPHOLD, AVKLAR_FORTSATT_MEDLEMSKAP];
 
-const buildInitialValues = createSelector(
-  [getFagsakPerson, getBehandlingMedlem, getPersonopplysning, behandlingSelectors.getSoknad, behandlingSelectors.getAksjonspunkter,
-    getBehandlingRevurderingAvFortsattMedlemskapFom, getAlleKodeverk],
-  (person, medlem, personopplysning, soknad, allAksjonspunkter, gjeldendeFom, alleKodeverk) => {
-    const aksjonspunkter = allAksjonspunkter
-      .filter((ap) => medlemAksjonspunkter.includes(ap.definisjon.kode))
-      .filter((ap) => ap.definisjon.kode !== aksjonspunktCodes.AVKLAR_STARTDATO_FOR_FORELDREPENGERPERIODEN);
-    let oppholdValues = {};
-    let confirmValues = {};
-    if (hasAksjonspunkt(AVKLAR_OPPHOLDSRETT, aksjonspunkter) || hasAksjonspunkt(AVKLAR_LOVLIG_OPPHOLD, aksjonspunkter)) {
-      oppholdValues = StatusForBorgerFaktaPanel.buildInitialValues(medlem, aksjonspunkter);
-    }
-    if (aksjonspunkter.length > 0) {
-      confirmValues = FaktaBegrunnelseTextField.buildInitialValues(aksjonspunkter[0]);
-    }
+const transformValues = (values, aksjonspunkter) => {
+  const aktiveMedlemAksjonspunkter = aksjonspunkter
+    .filter((ap) => medlemAksjonspunkter.includes(ap.definisjon.kode))
+    .filter((ap) => ap.definisjon.kode !== aksjonspunktCodes.AVKLAR_STARTDATO_FOR_FORELDREPENGERPERIODEN);
 
-    return {
-      ...OppholdINorgeOgAdresserFaktaPanel.buildInitialValues(soknad, personopplysning, medlem, aksjonspunkter),
-      ...InntektOgYtelserFaktaPanel.buildInitialValues(person, medlem),
-      ...PerioderMedMedlemskapFaktaPanel.buildInitialValues(medlem, soknad, aksjonspunkter, getKodeverknavnFn(alleKodeverk, kodeverkTyper)),
-      ...FortsattMedlemskapFaktaPanel.buildInitialValues(gjeldendeFom),
-      ...oppholdValues,
-      ...confirmValues,
-    };
-  },
-);
 
-const transformValues = (values, aksjonspunkter, state) => {
-  const aksjonspunkterArray = [];
-  if (shouldSubmitAksjonspunkt(AVKLAR_FORTSATT_MEDLEMSKAP, aksjonspunkter)) {
-    aksjonspunkterArray.push(FortsattMedlemskapFaktaPanel.transformValues(values));
-  }
-  if (shouldSubmitAksjonspunkt(AVKLAR_OM_BRUKER_ER_BOSATT, aksjonspunkter)) {
-    aksjonspunkterArray.push(OppholdINorgeOgAdresserFaktaPanel.transformValues(values));
-  }
-  if (shouldSubmitAksjonspunkt(AVKLAR_OM_BRUKER_HAR_GYLDIG_PERIODE, aksjonspunkter)) {
-    const vurderingstyper = getKodeverk(kodeverkTyper.MEDLEMSKAP_MANUELL_VURDERING_TYPE)(state);
-    aksjonspunkterArray.push(PerioderMedMedlemskapFaktaPanel.transformValues(values, vurderingstyper));
-  }
-  if (shouldSubmitAksjonspunkt(AVKLAR_OPPHOLDSRETT, aksjonspunkter) || shouldSubmitAksjonspunkt(AVKLAR_LOVLIG_OPPHOLD, aksjonspunkter)) {
-    aksjonspunkterArray.push(StatusForBorgerFaktaPanel.transformValues(values, aksjonspunkter));
-  }
-
-  const aksjonspunkterMedBegrunnelse = aksjonspunkterArray.map((ap) => ({
-    ...ap,
-    ...{ begrunnelse: values.begrunnelse },
+  return aktiveMedlemAksjonspunkter.map((ap) => ({
+    kode: ap.definisjon.kode,
+    begrunnelse: '',
+    bekreftedePerioder: values.perioder.map((periode) => {
+      // TODO Kor mange felt er det i bekreftetPeriode? Kan ein heller laga nytt objekt med det ein treng?
+      const {
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        id, fixedMedlemskapPerioder, foreldre, inntekter, manuellVurderingType, hasBosattAksjonspunkt, hasPeriodeAksjonspunkt,
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        isBosattAksjonspunktClosed, isPeriodAksjonspunktClosed, opphold, personopplysninger, fom, termindato, årsaker,
+        ...bekreftetPeriode
+      } = periode;
+      return bekreftetPeriode;
+    }).filter((periode) => periode.aksjonspunkter.includes(ap.definisjon.kode)
+      || (periode.aksjonspunkter.length > 0 && ap.definisjon.kode === aksjonspunktCodes.AVKLAR_FORTSATT_MEDLEMSKAP)),
   }));
-
-  return aksjonspunkterMedBegrunnelse;
 };
+
+
+const buildInitalValues = createSelector([behandlingSelectors.getSoknad, getFagsakPerson,
+  getBehandlingMedlemNew, getBehandlingRevurderingAvFortsattMedlemskapFom],
+(soknad, person, medlem = {}, gjeldendeFom = undefined) => ({
+  soknad,
+  person,
+  gjeldendeFom,
+  medlemskapPerioder: medlem.medlemskapPerioder || [],
+  inntekter: medlem.inntekt,
+  perioder: (medlem.perioder || []).map((periode) => ({
+    ...periode,
+    id: guid(),
+  })),
+}));
 
 const mapStateToPropsFactory = (initialState, ownProps) => {
-  const onSubmit = (values) => ownProps.submitCallback(transformValues(values, ownProps.aksjonspunkter, initialState));
+  const onSubmit = (values) => ownProps.submitCallback(transformValues(values, ownProps.aksjonspunkter));
   const hasOpenAksjonspunkter = ownProps.aksjonspunkter.some((ap) => isAksjonspunktOpen(ap.status.kode));
-  return (state) => ({
-    initialValues: buildInitialValues(state),
-    hasOpenAksjonspunkter,
-    onSubmit,
-  });
+  const perioder = [];
+
+  return (state) => {
+    const behandlingFormPrefix = getBehandlingFormPrefix(getSelectedBehandlingId(state), behandlingSelectors.getBehandlingVersjon(state));
+    return {
+      behandlingFormPrefix,
+      onSubmit,
+      hasOpenAksjonspunkter,
+      initialValues: buildInitalValues(state),
+      perioder: behandlingFormValueSelector('OppholdInntektOgPerioderForm')(state, 'perioder') || perioder,
+      isRevurdering: isBehandlingRevurderingFortsattMedlemskap(state),
+    };
+  };
 };
 
-export default connect(mapStateToPropsFactory)(behandlingFormForstegangOgRevurdering({
+const mapDispatchToProps = (dispatch) => ({
+  ...bindActionCreators({
+    reduxFormChange,
+    reduxFormReset,
+  }, dispatch),
+});
+
+export default connect(mapStateToPropsFactory, mapDispatchToProps)(behandlingFormForstegangOgRevurdering({
   form: 'OppholdInntektOgPerioderForm',
-})(OppholdInntektOgPerioderForm));
+  enableReinitialize: true,
+})(injectIntl(OppholdInntektOgPerioderFormNew)));
