+++ packages/fp-behandling-forstegang-og-revurdering/src/fakta/components/medlemskap/oppholdInntektOgPerioder/PerioderMedMedlemskapFaktaPanel.jsx
@@ -2,22 +2,21 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { connect } from 'react-redux';
-import { Column, Row } from 'nav-frontend-grid';
 import { Normaltekst } from 'nav-frontend-typografi';
 import moment from 'moment';
 
-import { behandlingFormValueSelector } from 'behandlingForstegangOgRevurdering/src/behandlingFormForstegangOgRevurdering';
 import { RadioGroupField, RadioOption } from '@fpsak-frontend/form';
-import FaktaGruppe from 'behandlingForstegangOgRevurdering/src/fakta/components/FaktaGruppe';
-
 import {
-  DateLabel, PeriodLabel, Table, TableColumn, TableRow, VerticalSpacer,
+  DateLabel, FlexColumn, FlexContainer, FlexRow, PeriodLabel, Table, TableColumn, TableRow, VerticalSpacer,
 } from '@fpsak-frontend/shared-components';
 import aksjonspunktCodes from '@fpsak-frontend/kodeverk/src/aksjonspunktCodes';
-import { getKodeverk } from 'behandlingForstegangOgRevurdering/src/duckBehandlingForstegangOgRev';
 import kodeverkTyper from '@fpsak-frontend/kodeverk/src/kodeverkTyper';
-import { isAksjonspunktOpen } from '@fpsak-frontend/kodeverk/src/aksjonspunktStatus';
 import { DDMMYYYY_DATE_FORMAT, required } from '@fpsak-frontend/utils';
+import { isAksjonspunktOpen } from '@fpsak-frontend/kodeverk/src/aksjonspunktStatus';
+
+import { getKodeverk } from 'behandlingForstegangOgRevurdering/src/duckBehandlingForstegangOgRev';
+import { behandlingFormValueSelector } from 'behandlingForstegangOgRevurdering/src/behandlingFormForstegangOgRevurdering';
+import FaktaGruppe from 'behandlingForstegangOgRevurdering/src/fakta/components/FaktaGruppe';
 
 const headerTextCodes = [
   'PerioderMedMedlemskapFaktaPanel.Period',
@@ -35,13 +34,13 @@ export const PerioderMedMedlemskapFaktaPanelImpl = ({
   readOnly,
   hasPeriodeAksjonspunkt,
   isPeriodAksjonspunktClosed,
-  periods,
+  fixedMedlemskapPerioder,
   fodselsdato,
   termindato,
   omsorgsovertakelseDato,
   vurderingTypes,
 }) => {
-  if (!periods || periods.length === 0) {
+  if (!fixedMedlemskapPerioder || fixedMedlemskapPerioder.length === 0) {
     return (
       <FaktaGruppe titleCode="PerioderMedMedlemskapFaktaPanel.ApplicationInformation">
         <Normaltekst>
@@ -54,70 +53,71 @@ export const PerioderMedMedlemskapFaktaPanelImpl = ({
   return (
     <FaktaGruppe aksjonspunktCode={aksjonspunktCodes.AVKLAR_OM_BRUKER_HAR_GYLDIG_PERIODE} titleCode="PerioderMedMedlemskapFaktaPanel.ApplicationInformation">
       <Table headerTextCodes={headerTextCodes}>
-        {periods.map((period) => {
-          const key = period.fom + period.tom + period.dekning + period.status + period.beslutningsdato;
+        {fixedMedlemskapPerioder.map((periode) => {
+          const key = periode.fom + periode.tom + periode.dekning + periode.status + periode.beslutningsdato;
           return (
             <TableRow key={key} id={key}>
               <TableColumn>
-                <PeriodLabel showTodayString dateStringFom={period.fom} dateStringTom={period.tom} />
+                <PeriodLabel showTodayString dateStringFom={periode.fom} dateStringTom={periode.tom} />
               </TableColumn>
               <TableColumn>
-                {period.dekning}
+                {periode.dekning}
               </TableColumn>
               <TableColumn>
-                {period.status}
+                {periode.status}
               </TableColumn>
               <TableColumn>
-                {period.beslutningsdato ? <DateLabel dateString={period.beslutningsdato} /> : null}
+                {periode.beslutningsdato ? <DateLabel dateString={periode.beslutningsdato} /> : null}
               </TableColumn>
             </TableRow>
           );
         })}
       </Table>
-      {hasPeriodeAksjonspunkt
+      <FlexContainer>
+        {hasPeriodeAksjonspunkt
       && (
-      <Row>
-        <Column xs="12">
-          <RadioGroupField name="manuellVurderingType.kode" validate={[required]} readOnly={readOnly} isEdited={isPeriodAksjonspunktClosed}>
+      <FlexRow>
+        <FlexColumn>
+          <RadioGroupField name="medlemskapManuellVurderingType.kode" validate={[required]} readOnly={readOnly} isEdited={isPeriodAksjonspunktClosed}>
             {vurderingTypes.map((type) => <RadioOption key={type.kode} value={type.kode} label={type.navn} />)}
           </RadioGroupField>
-        </Column>
-      </Row>
+        </FlexColumn>
+      </FlexRow>
       )}
-      <VerticalSpacer sixteenPx />
-      <Row>
-        <Column xs="8" />
-        <Column xs="4">
-          {fodselsdato
+        <VerticalSpacer sixteenPx />
+        <FlexRow className="justifyItemsToFlexEnd">
+          <FlexColumn>
+            {fodselsdato
           && (
           <FormattedMessage
             id="PerioderMedMedlemskapFaktaPanel.Fodselsdato"
             values={{ dato: moment(fodselsdato).format(DDMMYYYY_DATE_FORMAT) }}
           />
           )}
-          {termindato
+            {termindato
           && (
           <FormattedMessage
             id="PerioderMedMedlemskapFaktaPanel.Termindato"
             values={{ dato: moment(termindato).format(DDMMYYYY_DATE_FORMAT) }}
           />
           )}
-          {omsorgsovertakelseDato
+            {omsorgsovertakelseDato
           && (
           <FormattedMessage
             id="PerioderMedMedlemskapFaktaPanel.Omsorgsovertakelse"
             values={{ dato: moment(omsorgsovertakelseDato).format(DDMMYYYY_DATE_FORMAT) }}
           />
           )}
-        </Column>
-      </Row>
+          </FlexColumn>
+        </FlexRow>
+      </FlexContainer>
     </FaktaGruppe>
   );
 };
 
 PerioderMedMedlemskapFaktaPanelImpl.propTypes = {
   readOnly: PropTypes.bool.isRequired,
-  periods: PropTypes.arrayOf(PropTypes.shape()),
+  fixedMedlemskapPerioder: PropTypes.arrayOf(PropTypes.shape()),
   fodselsdato: PropTypes.string,
   termindato: PropTypes.string,
   omsorgsovertakelseDato: PropTypes.string,
@@ -130,12 +130,12 @@ PerioderMedMedlemskapFaktaPanelImpl.defaultProps = {
   fodselsdato: undefined,
   termindato: undefined,
   omsorgsovertakelseDato: undefined,
-  periods: [],
+  fixedMedlemskapPerioder: [],
 };
 
-const mapStateToProps = (state) => ({
-  ...behandlingFormValueSelector('OppholdInntektOgPerioderForm')(
-    state, 'periods', 'fodselsdato', 'termindato',
+const mapStateToProps = (state, ownProps) => ({
+  ...behandlingFormValueSelector(`OppholdInntektOgPeriodeForm-${ownProps.id}`)(
+    state, 'fixedMedlemskapPerioder', 'fodselsdato', 'termindato',
     'omsorgsovertakelseDato', 'hasPeriodeAksjonspunkt', 'isPeriodAksjonspunktClosed',
   ),
   vurderingTypes: getKodeverk(kodeverkTyper.MEDLEMSKAP_MANUELL_VURDERING_TYPE)(state),
@@ -143,25 +143,27 @@ const mapStateToProps = (state) => ({
 
 const PerioderMedMedlemskapFaktaPanel = connect(mapStateToProps)(PerioderMedMedlemskapFaktaPanelImpl);
 
-PerioderMedMedlemskapFaktaPanel.buildInitialValues = (medlem, soknad, aksjonspunkter, getKodeverknavn) => {
-  if (medlem === null) {
+PerioderMedMedlemskapFaktaPanel.buildInitialValues = (periode, medlemskapPerioder, soknad, aksjonspunkter, getKodeverknavn) => {
+  if (medlemskapPerioder === null) {
     return [];
   }
 
-  const periods = medlem.medlemskapPerioder
-    .map((i) => ({
-      fom: i.fom,
-      tom: i.tom,
-      dekning: getKodeverknavn(i.dekningType),
-      status: getKodeverknavn(i.medlemskapType),
-      beslutningsdato: i.beslutningsdato,
-    }))
+  const fixedMedlemskapPerioder = medlemskapPerioder.map((i) => ({
+    fom: i.fom,
+    tom: i.tom,
+    dekning: getKodeverknavn(i.dekningType),
+    status: getKodeverknavn(i.medlemskapType),
+    beslutningsdato: i.beslutningsdato,
+  }))
     .sort((p1, p2) => new Date(p1.fom).getTime() - new Date(p2.fom).getTime());
+  const filteredAp = aksjonspunkter.filter((ap) => periode.aksjonspunkter.includes(ap.definisjon.kode)
+      || (periode.aksjonspunkter.length > 0
+        && periode.aksjonspunkter.includes(aksjonspunktCodes.AVKLAR_OM_BRUKER_HAR_GYLDIG_PERIODE)
+        && ap.definisjon.kode === aksjonspunktCodes.AVKLAR_FORTSATT_MEDLEMSKAP));
 
-  const filteredAp = aksjonspunkter.filter((ap) => ap.definisjon.kode === aksjonspunktCodes.AVKLAR_OM_BRUKER_HAR_GYLDIG_PERIODE);
   return {
-    periods,
-    manuellVurderingType: medlem.medlemskapManuellVurderingType,
+    fixedMedlemskapPerioder,
+    medlemskapManuellVurderingType: periode.medlemskapManuellVurderingType,
     fodselsdato: soknad.fodselsdatoer ? Object.values(soknad.fodselsdatoer)[0] : undefined,
     termindato: soknad.termindato,
     omsorgsovertakelseDato: soknad.omsorgsovertakelseDato,
@@ -172,7 +174,7 @@ PerioderMedMedlemskapFaktaPanel.buildInitialValues = (medlem, soknad, aksjonspun
 
 PerioderMedMedlemskapFaktaPanel.transformValues = (values, manuellVurderingTyper) => ({
   kode: aksjonspunktCodes.AVKLAR_OM_BRUKER_HAR_GYLDIG_PERIODE,
-  manuellVurderingType: manuellVurderingTyper.find((m) => m.kode === values.manuellVurderingType.kode),
+  medlemskapManuellVurderingType: manuellVurderingTyper.find((m) => m.kode === values.medlemskapManuellVurderingType.kode),
 });
 
 export default PerioderMedMedlemskapFaktaPanel;
