--- packages/fp-behandling-forstegang-og-revurdering/src/fakta/components/medlemskap/oppholdInntektOgPerioder/statusForBorgerFaktaPanel.less
@@ -1,23 +1,0 @@
-@import '@{coreModulePath}nav-frontend-core/less/_variabler';
-
-.error {
-  padding-left: 0;
-  padding-right: 15px;
-}
-
-.inline {
-  display: inline;
-}
-
-.radioGroup {
-  display: inline-block;
-  width: 550px;
-}
-
-.radioGroupLeft {
-  float: left;
-}
-
-.radioGroupRight {
-  float: right;
-}
