--- packages/fp-behandling-forstegang-og-revurdering/src/fakta/components/medlemskap/oppholdInntektOgPerioderNew/oppholdINorgeOgAdresserFaktaPanel.less
@@ -1,14 +1,0 @@
-@import '@{coreModulePath}nav-frontend-core/less/_variabler';
-
-.dodEtikett {
-  background-color: @navMorkGra;
-  color: @white;
-}
-
-@media screen and (-ms-high-contrast: active), screen and (-ms-high-contrast: none) {
-  .ieFlex {
-    [class^='flexRow'] {
-      min-width: 300px;
-    }
-  }
-}
