+++ packages/fp-felles/src/featureToggle.jsx
@@ -1,5 +1,4 @@
 const featureToggle = {
-  LØPENDE_MEDLESMKAP: 'fpsak.lopende-medlemskap',
   AKTIVER_TILBAKEKREVINGBEHANDLING: 'fpsak.aktiver-tilbakekrevingbehandling',
   SIMULER_VARSELTEKST: 'fpsak.simuler-oppdrag-varseltekst',
   MARKER_UTENLANDSSAK: 'fpsak.aksjonspunkt-marker-utenlandssak',
